src/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
fastapi==0.115.8
idna==3.10
jose==1.0.0
passlib==1.7.4
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.45.3
typing_extensions==4.12.2
